+++ .github/workflows/ci.yml
@@ -40,7 +40,8 @@ jobs:
         python-version: 3.7
     - name: Install Dependencies
       run: |
-        pip install black==20.8b1
+        pip install click==8.0.4
+        pip install black==21.6b0
     - name: Run Black
       run: |
         black --check sleap tests
