+++ dev_requirements.txt
@@ -19,3 +19,4 @@ PyGithub
 jupyterlab
 jedi==0.17.2
 ipykernel
+click==8.0.4
